requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
bitsandbytes>=0.43.3   # 只指定最低版本，避免版本冲突
protobuf<4.0.0
hf_transfer>=0.1.8     # 放宽版本限制
trl>=0.9.4
peft>=0.11.1
accelerate>=0.31.0     # 增加 accelerate
flask>=3.0.3
nacos-sdk-python>=0.1.14
eth-utils>=4.1.1
xformers>=0.0.26.post1
